requirement.txt
langchain==0.2.3
langchain_community
langchain_chroma
langchain_experimental
langchain_google_genai
python-dotenv
pypdf==3.12.0
sentence_transformers
streamlit==1.35.0
streamlit_chat
chromadb